+++ protocols/gossipsub/src/handler.rs
@@ -22,6 +22,7 @@ use crate::protocol::{GossipsubCodec, ProtocolConfig};
 use crate::rpc_proto::proto;
 use crate::types::{PeerKind, RawMessage, Rpc};
 use crate::ValidationError;
+use arraydeque::{ArrayDeque, Wrapping};
 use asynchronous_codec::Framed;
 use futures::future::Either;
 use futures::prelude::*;
@@ -33,7 +34,6 @@ use libp2p_swarm::handler::{
     FullyNegotiatedInbound, FullyNegotiatedOutbound, StreamUpgradeError, SubstreamProtocol,
 };
 use libp2p_swarm::Stream;
-use smallvec::SmallVec;
 use std::{
     pin::Pin,
     task::{Context, Poll},
@@ -95,7 +95,7 @@ pub struct EnabledHandler {
     inbound_substream: Option<InboundSubstreamState>,
 
     /// Queue of values that we want to send to the remote.
-    send_queue: SmallVec<[proto::RPC; 16]>,
+    send_queue: ArrayDeque<proto::RPC, 16, Wrapping>,
 
     /// Flag indicating that an outbound substream is being established to prevent duplicate
     /// requests.
@@ -167,7 +167,7 @@ impl Handler {
             outbound_substream_establishing: false,
             outbound_substream_attempts: 0,
             inbound_substream_attempts: 0,
-            send_queue: SmallVec::new(),
+            send_queue: ArrayDeque::new(),
             peer_kind: None,
             peer_kind_sent: false,
             last_io_activity: Instant::now(),
@@ -315,8 +315,7 @@ impl EnabledHandler {
             ) {
                 // outbound idle state
                 Some(OutboundSubstreamState::WaitingOutput(substream)) => {
-                    if let Some(message) = self.send_queue.pop() {
-                        self.send_queue.shrink_to_fit();
+                    if let Some(message) = self.send_queue.pop_front() {
                         self.outbound_substream =
                             Some(OutboundSubstreamState::PendingSend(substream, message));
                         continue;
@@ -409,7 +408,7 @@ impl ConnectionHandler for Handler {
     fn on_behaviour_event(&mut self, message: HandlerIn) {
         match self {
             Handler::Enabled(handler) => match message {
-                HandlerIn::Message(m) => handler.send_queue.push(m),
+                HandlerIn::Message(m) => _ = handler.send_queue.push_back(m), // drops frontmost element if full
                 HandlerIn::JoinedMesh => {
                     handler.in_mesh = true;
                 }